Moved examples to grammars folder and linked in web
