+++ grammars/json.stripped.json
