+++ js/web/src/components/File.tsx
@@ -46,6 +46,8 @@ function ExampleSelect({
     value: LanguageKey | undefined;
     onChange: (pair: PlaygroundPair) => void;
 }) {
+    console.log(playgroundPairs);
+
     return (
         <select
             value={value}
