--- js/web/src/examples/jsonStripped.peg
@@ -1,9 +1,0 @@
-JSON    <- Value EOF
-Value   <- Object / Array / String / Number / 'true' / 'false' / 'null'
-Array   <- '[' (Value (',' Value)*)? ']'
-Object  <- '{' (Member (',' Member)*)? '}'
-Member  <- String ':' Value
-Number  <- '-'? ('0' / ([1-9][0-9]*)) ('.' [0-9]+)? ([eE][-+]?[0-9]+)?
-String  <- '"' (!["\\] .
-              / ('\\' (["\\/bfnrt] / 'u' [0-9A-Fa-f][0-9A-Fa-f][0-9A-Fa-f][0-9A-Fa-f])))*
-           '"'
