--- js/web/src/examples/tiny.peg
@@ -1,9 +1,0 @@
-Expr     <- Multi (( '+' / '-' ) Multi)*
-Multi    <- Primary (( '*' / '/' ) Primary)*
-Primary  <- Call / Id / Num / '(' Expr ')'
-
-Call     <- Id '(' Params ')'
-Params   <- (Expr (',' Expr)*)?
-
-Num      <- [1-9][0-9]* / '0'
-Id       <- [a-zA-Z_][a-zA-Z0-9_]*
