--- js/web/src/examples/tiny.tiny
@@ -1,1 +1,0 @@
-he+1
