--- js/web/src/examples/xmlUnstable.peg
@@ -1,12 +1,0 @@
-TopLevel  <- PROLOG? _* DTD? _* Element _*
-PROLOG    <- '<?xml' (!'?>' .)* '?>'
-DTD       <- '<' (!'>' .)* '>'
-Element   <- '<' Name (_+ Attribute)* ('/>' / '>' Content '</' Name '>') _*
-Name      <- [A-Za-z:] ('-' / [A-Za-z0-9:._])*
-Attribute <- Name _* '=' _* String
-String    <- '"' (!'"' .)* '"'
-Content   <- (Element / CDataSec / CharData)*
-CDataSec  <- '<![CDATA[' (!']]>' .)* ']]>' _*
-COMMENT   <- '<!--' (!'-->' .)* '-->' _*
-CharData  <- (!'<' .)+
-_         <- [ \t\r\n]
