+++ js/web/src/grammars
@@ -1,0 +1,1 @@
+../../../grammars
